Login with no authentication

+++ app/src/main/java/com/gyonia/exercisetracker/ui/login/LoginViewModel.kt
@@ -22,7 +22,7 @@ class LoginViewModel(private val loginRepository: LoginRepository) : ViewModel()
 
         if (result is Result.Success) {
             _loginResult.value =
-                LoginResult(success = LoggedInUserView(displayName = result.data.displayName, userId = result.data.userId.toInt()))
+                LoginResult(success = LoggedInUserView(displayName = result.data.displayName, userId = result.data.userId))
         } else {
             _loginResult.value = LoginResult(error = R.string.login_failed)
         }
